Code cleanup

+++ utils/utils.cpp
@@ -5,8 +5,8 @@
 std::string Utils::matType2Str(int type) {
     std::string r;
 
-    uchar depth = type & CV_MAT_DEPTH_MASK;
-    uchar chans = 1 + (type >> CV_CN_SHIFT);
+    uchar depth = static_cast<uchar>(type & CV_MAT_DEPTH_MASK);
+    uchar chans = static_cast<uchar>(1 + (type >> CV_CN_SHIFT));
 
     switch (depth) {
         case CV_8U:  r = "8U"; break;
